refactor(ps): Prefer per-platform Snapshotter interfaces
Drop the shared Resolver composition. Linux Snapshotter uses uint64 PIDs, Windows keeps uint32, and event.PID becomes a platform type alias so shared rule/filter call sites stay typed correctly.

+++ pkg/config/filters.go
@@ -176,7 +176,7 @@ func (ctx *ActionContext) UniquePids() []uint32 {
 		if e.IsCreateProcess() {
 			pids[e.Params.MustGetPid()] = struct{}{}
 		} else {
-			pids[e.PID] = struct{}{}
+			pids[uint32(e.PID)] = struct{}{}
 		}
 	}
 	return convert.MapKeysToSlice(pids)
